backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
